refactor: move SQL-specific code to laboratory fix typos
