+++ packages/laboratory/package.json
@@ -1,6 +1,9 @@
 {
   "name": "@microsoft/sds-laboratory",
   "version": "0.1.0",
+  "repository": {
+    "url": "https://github.com/microsoft/secure-data-sandbox"
+  },
   "description": "A toolkit for conducting machine learning trials against confidential data",
   "license": "MIT",
   "main": "dist/index",
@@ -26,27 +29,40 @@
     "test": "ts-mocha"
   },
   "devDependencies": {
+    "@types/bluebird": "^3.5.32",
     "@types/chai": "^4.2.12",
+    "@types/chai-as-promised": "^7.1.3",
     "@types/chai-http": "^4.2.0",
     "@types/express": "^4.17.8",
     "@types/mocha": "^8.0.3",
     "@types/node": "^13.13.20",
     "@types/passport-azure-ad": "^4.0.7",
+    "@types/tedious": "^4.0.1",
+    "@types/uuid": "^8.3.0",
+    "@types/validator": "^13.1.0",
     "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "chai-exclude": "^2.0.2",
     "chai-http": "^4.3.0",
     "gts": "^2.0.2",
     "mocha": "^8.1.3",
     "ts-mocha": "^7.0.0",
-    "typescript": "^3.8.3",
-    "xmlhttprequest": "^1.8.0"
+    "typescript": "^3.8.3"
   },
   "dependencies": {
+    "@azure/identity": "^1.1.0",
     "@microsoft/sds": "*",
     "env-var": "^6.3.0",
     "express": "^4.17.1",
     "express-async-errors": "^3.1.1",
     "passport": "^0.4.1",
     "passport-azure-ad": "^4.3.0",
-    "strong-error-handler": "^3.5.0"
+    "reflect-metadata": "^0.1.13",
+    "sequelize": "^5.22.3",
+    "sequelize-typescript": "^1.1.0",
+    "sqlite3": "^5.0.0",
+    "strong-error-handler": "^3.5.0",
+    "tedious": "^9.2.1",
+    "uuid": "^8.3.0"
   }
 }
