+++ packages/laboratory/src/sequelize_laboratory/index.ts
@@ -1,4 +1,3 @@
 export * from './laboratory';
-export * from './normalize';
 export * from './sequelize';
 // models are private to sequelize_laboratory
