+++ packages/laboratory/src/sequelize_laboratory/models/run.ts
@@ -6,7 +6,7 @@ import {
   IRun,
   ISuite,
   RunStatus,
-} from '../../interfaces';
+} from '@microsoft/sds';
 
 import { jsonColumn } from './decorators';
