+++ packages/laboratory/src/sequelize_laboratory/models/suite.ts
@@ -1,11 +1,11 @@
 import { DataType, Column, Model, Table } from 'sequelize-typescript';
 
-import { IBenchmark, PipelineStage } from '../../interfaces';
+import { ISuite, SuiteVolume } from '@microsoft/sds';
 
 import { jsonColumn } from './decorators';
 
 @Table
-export class Benchmark extends Model<Benchmark> implements IBenchmark {
+export class Suite extends Model<Suite> implements ISuite {
   @Column({
     type: DataType.STRING,
     primaryKey: true,
@@ -18,7 +18,10 @@ export class Benchmark extends Model<Benchmark> implements IBenchmark {
   @Column(DataType.STRING)
   apiVersion!: string;
 
+  @Column(DataType.STRING)
+  benchmark!: string;
+
   // TODO: REVIEW: magic number 1024
-  @Column(jsonColumn<PipelineStage[]>('stages', 1024))
-  stages!: PipelineStage[];
+  @Column(jsonColumn<SuiteVolume[]>('volumes', 1024))
+  volumes!: SuiteVolume[];
 }
