+++ packages/laboratory/tsconfig.build.json
@@ -1,6 +1,7 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
+    "experimentalDecorators": true,
     "outDir": "dist",
     "rootDir": "src",
   },
