+++ packages/laboratory/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
+    "experimentalDecorators": true,
     "outDir": "dist",
     "baseUrl": "../",
     "paths": {
