+++ packages/sds/package.json
@@ -1,6 +1,9 @@
 {
   "name": "@microsoft/sds",
   "version": "0.1.0",
+  "repository": {
+    "url": "https://github.com/microsoft/secure-data-sandbox"
+  },
   "description": "A toolkit for conducting machine learning trials against confidential data",
   "license": "MIT",
   "main": "dist/index",
@@ -23,28 +26,23 @@
   },
   "devDependencies": {
     "@sinonjs/fake-timers": "^6.0.1",
-    "@types/bluebird": "^3.5.32",
     "@types/chai": "^4.2.12",
     "@types/chai-as-promised": "^7.1.3",
-    "@types/dotenv": "^8.2.0",
     "@types/luxon": "^1.24.4",
     "@types/mocha": "^8.0.3",
     "@types/nock": "^11.1.0",
     "@types/node": "^13.13.20",
     "@types/sinonjs__fake-timers": "^6.0.1",
-    "@types/tedious": "^4.0.1",
     "@types/uuid": "^8.3.0",
-    "@types/validator": "^13.1.0",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "chai-exclude": "^2.0.2",
     "dotenv": "^8.2.0",
     "gts": "^2.0.2",
     "mocha": "^8.1.3",
     "nock": "^13.0.4",
     "ts-mocha": "^7.0.0",
     "typescript": "^3.8.3",
-    "xmlhttprequest": "^1.8.0"
+    "uuid": "^8.3.0"
   },
   "dependencies": {
     "@azure/identity": "^1.1.0",
@@ -53,12 +51,6 @@
     "env-var": "^6.3.0",
     "fp-ts": "^2.8.2",
     "io-ts": "^2.2.10",
-    "luxon": "^1.25.0",
-    "reflect-metadata": "^0.1.13",
-    "sequelize": "^5.22.3",
-    "sequelize-typescript": "^1.1.0",
-    "sqlite3": "^5.0.0",
-    "tedious": "^9.2.1",
-    "uuid": "^8.3.0"
+    "luxon": "^1.25.0"
   }
 }
