+++ packages/sds/src/index.ts
@@ -1,5 +1,4 @@
 export * from './laboratory';
 export * from './queue';
 export * from './messages';
-export * from './database';
 export * as configuration from './configuration';
