--- packages/sds/src/laboratory/logic/index.ts
@@ -1,3 +1,0 @@
-export * from './interfaces';
-export * from './sequelize_laboratory';
-export * from './validate';
